+++ CHANGELOG.md
@@ -15,6 +15,7 @@ See [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for more details.
 - record.py file
 - seismicsource.py file
 - station.py file
+- hercules incident folder arrangment
 
 ### Changed
 - package name (ts-process -> tsprocess)